Add unit tests, starting with the Item api methods
Running the tests requires mock and nose. To install them:

    pip install mock nose

To run the tests, run `nosetests` from the repository root.

I avoided touching the pypodio2 files where possible, but areas.py was
failing because the docstring for Item.update didn't have the same
indentation as the rest of the method. I didn't make any functional
changes to the API client itself.