Update to areas.Task methods
Adds a `find` method on the `Task` area to retrieve a single task by the id. This has been named `Task.find` to remain consistent with other areas.  The `get` method is not consistent with other areas though which use `find_all` or `find_all_by_*`. So `find_all` has been added and a comment placed in `get` to deprecate this method.